Move playground modules into ui package
