+++ playground/src/main/java/mx/yellowme/youst/playground/ui/navigation/NavigationActivity.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.playground.navigation
+package mx.yellowme.youst.playground.ui.navigation
 
 import mx.yellowme.youst.playground.R
 import mx.yellowme.youst.core.hooks.BaseActivity
