+++ playground/src/main/java/mx/yellowme/youst/playground/ui/navigation/ScreenAFragment.kt
@@ -1,19 +1,20 @@
-package mx.yellowme.youst.playground.navigation
+package mx.yellowme.youst.playground.ui.navigation
 
 import androidx.navigation.fragment.findNavController
-import kotlinx.android.synthetic.main.screen_user_no_attribute.*
+import kotlinx.android.synthetic.main.screen_login.*
 import mx.yellowme.youst.core.hooks.BaseFragment
 import mx.yellowme.youst.playground.R
 import mx.yellowme.youst.playground.components.AppHeroActionListener
 
-class ScreenEFragment : BaseFragment() {
+class ScreenAFragment : BaseFragment() {
 
-    override val layoutId: Int = R.layout.screen_e
+    override val layoutId: Int = R.layout.screen_a
 
     override fun onViewReady() {
         appHero.mainActionListener = object : AppHeroActionListener {
             override fun onClickAction() {
-                val action = ScreenEFragmentDirections.actionScreenEToScreenF()
+                val action =
+                    ScreenAFragmentDirections.actionScreenAToScreenB()
                 findNavController().navigate(action)
             }
         }
