+++ playground/src/main/java/mx/yellowme/youst/playground/ui/navigation/ScreenBFragment.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.playground.navigation
+package mx.yellowme.youst.playground.ui.navigation
 
 import androidx.navigation.fragment.findNavController
 import kotlinx.android.synthetic.main.screen_login.*
@@ -6,9 +6,9 @@ import mx.yellowme.youst.core.hooks.BaseFragment
 import mx.yellowme.youst.playground.R
 import mx.yellowme.youst.playground.components.AppHeroActionListener
 
-class ScreenCFragment : BaseFragment() {
+class ScreenBFragment : BaseFragment() {
 
-    override val layoutId: Int = R.layout.screen_c
+    override val layoutId: Int = R.layout.screen_b
 
     override fun onViewReady() {
 
@@ -16,14 +16,16 @@ class ScreenCFragment : BaseFragment() {
 
             mainActionListener = object : AppHeroActionListener {
                 override fun onClickAction() {
-                    val action = ScreenCFragmentDirections.actionScreenCToScreenG()
+                    val action =
+                        ScreenBFragmentDirections.actionScreenBToScreenC()
                     findNavController().navigate(action)
                 }
             }
 
             secondaryActionListener = object : AppHeroActionListener {
                 override fun onClickAction() {
-                    val action = ScreenCFragmentDirections.actionScreenCToScreenA()
+                    val action =
+                        ScreenBFragmentDirections.actionScreenBToScreenD()
                     findNavController().navigate(action)
                 }
             }
@@ -31,5 +33,4 @@ class ScreenCFragment : BaseFragment() {
         }
 
     }
-
 }
