+++ playground/src/main/java/mx/yellowme/youst/playground/ui/navigation/ScreenCFragment.kt
@@ -1,4 +1,4 @@
-package mx.yellowme.youst.playground.navigation
+package mx.yellowme.youst.playground.ui.navigation
 
 import androidx.navigation.fragment.findNavController
 import kotlinx.android.synthetic.main.screen_login.*
@@ -6,9 +6,9 @@ import mx.yellowme.youst.core.hooks.BaseFragment
 import mx.yellowme.youst.playground.R
 import mx.yellowme.youst.playground.components.AppHeroActionListener
 
-class ScreenBFragment : BaseFragment() {
+class ScreenCFragment : BaseFragment() {
 
-    override val layoutId: Int = R.layout.screen_b
+    override val layoutId: Int = R.layout.screen_c
 
     override fun onViewReady() {
 
@@ -16,14 +16,16 @@ class ScreenBFragment : BaseFragment() {
 
             mainActionListener = object : AppHeroActionListener {
                 override fun onClickAction() {
-                    val action = ScreenBFragmentDirections.actionScreenBToScreenC()
+                    val action =
+                        ScreenCFragmentDirections.actionScreenCToScreenG()
                     findNavController().navigate(action)
                 }
             }
 
             secondaryActionListener = object : AppHeroActionListener {
                 override fun onClickAction() {
-                    val action = ScreenBFragmentDirections.actionScreenBToScreenD()
+                    val action =
+                        ScreenCFragmentDirections.actionScreenCToScreenA()
                     findNavController().navigate(action)
                 }
             }
@@ -31,4 +33,5 @@ class ScreenBFragment : BaseFragment() {
         }
 
     }
+
 }
