+++ playground/src/main/java/mx/yellowme/youst/playground/ui/navigation/ScreenDFragment.kt
@@ -1,19 +1,20 @@
-package mx.yellowme.youst.playground.navigation
+package mx.yellowme.youst.playground.ui.navigation
 
 import androidx.navigation.fragment.findNavController
-import kotlinx.android.synthetic.main.screen_user_no_attribute.*
+import kotlinx.android.synthetic.main.screen_login.*
 import mx.yellowme.youst.core.hooks.BaseFragment
 import mx.yellowme.youst.playground.R
 import mx.yellowme.youst.playground.components.AppHeroActionListener
 
-class ScreenFFragment : BaseFragment() {
+class ScreenDFragment : BaseFragment() {
 
-    override val layoutId: Int = R.layout.screen_f
+    override val layoutId: Int = R.layout.screen_d
 
     override fun onViewReady() {
         appHero.mainActionListener = object : AppHeroActionListener {
             override fun onClickAction() {
-                val action = ScreenFFragmentDirections.actionScreenFToScreenB()
+                val action =
+                    ScreenDFragmentDirections.actionScreenDToScreenE()
                 findNavController().navigate(action)
             }
         }
