+++ playground/src/main/java/mx/yellowme/youst/playground/ui/navigation/ScreenEFragment.kt
@@ -1,19 +1,20 @@
-package mx.yellowme.youst.playground.navigation
+package mx.yellowme.youst.playground.ui.navigation
 
 import androidx.navigation.fragment.findNavController
-import kotlinx.android.synthetic.main.screen_login.*
+import kotlinx.android.synthetic.main.screen_user_no_attribute.*
 import mx.yellowme.youst.core.hooks.BaseFragment
 import mx.yellowme.youst.playground.R
 import mx.yellowme.youst.playground.components.AppHeroActionListener
 
-class ScreenAFragment : BaseFragment() {
+class ScreenEFragment : BaseFragment() {
 
-    override val layoutId: Int = R.layout.screen_a
+    override val layoutId: Int = R.layout.screen_e
 
     override fun onViewReady() {
         appHero.mainActionListener = object : AppHeroActionListener {
             override fun onClickAction() {
-                val action = ScreenAFragmentDirections.actionScreenAToScreenB()
+                val action =
+                    ScreenEFragmentDirections.actionScreenEToScreenF()
                 findNavController().navigate(action)
             }
         }
