+++ playground/src/main/java/mx/yellowme/youst/playground/ui/nemo/navigator/fetchers/FetchConfig.kt
@@ -1,8 +1,8 @@
-package mx.yellowme.youst.playground.nemo.navigator.fetchers
+package mx.yellowme.youst.playground.ui.nemo.navigator.fetchers
 
-import mx.yellowme.youst.playground.nemo.navigator.hooks.AsyncSingleNavigator
-import mx.yellowme.youst.playground.nemo.navigator.hooks.Navigator
-import mx.yellowme.youst.playground.nemo.navigator.BaseHelm
+import mx.yellowme.youst.playground.ui.nemo.navigator.hooks.AsyncSingleNavigator
+import mx.yellowme.youst.playground.ui.nemo.navigator.hooks.Navigator
+import mx.yellowme.youst.playground.ui.nemo.navigator.BaseHelm
 import mx.yellowme.youst.playground.data.ConfigFakeRepository
 import mx.yellowme.youst.playground.domain.Config
